Removed unnecessary code and corrected Kapacitor interface
