+++ src/interfaces/Alert.ts
@@ -1,41 +1,34 @@
-export default interface Alert {
-    _id: string;
-    alertId: string;
-    message: string;
-    details: {
-        Name: string;
-        TaskName: string;
-        Group: string;
-        Tags: {
-            [key: string]: string;
-        };
-        ServerInfo: {
-            Hostname: string;
-            ClusterID: string;
-            ServerID: string;
-        };
-        ID: string;
-        Fields: {
-            [key: string]: number;
-        };
-        Level: string;
-        Time: string;
-        Duration: number
-        Message: string;
+export default interface KapacitorAlert {
+    link: {
+        rel: string;
+        href: string;
     };
-    time: string;
-    duration: number;
-    level: string;
-    data: {
-        series: {
-            name: string;
-            tags: {
+    id: string;
+    state: {
+        message: string,
+        details: {
+            Name: string;
+            TaskName: string;
+            Group: string;
+            Tags: {
                 [key: string]: string;
             };
-            columns: string[];
-            values: string[];
+            ServerInfo: {
+                Hostname: string;
+                ClusterID: string;
+                ServerID: string;
+            };
+            ID: string;
+            Fields: {
+                [key: string]: number;
+            };
+            Level: string;
+            Time: string;
+            Duration: number
+            Message: string;
         };
+        time: string;
+        duration: string;
+        level: string;
     };
-    recoverable: boolean;
-    __v: number;
 }
