+++ src/views/EventList.vue
@@ -26,7 +26,7 @@
 import Vue from 'vue';
 import Component from 'vue-class-component';
 import Header from '@/components/Header.vue';
-import KapacitorAlert from '@/interfaces/KapacitorAlert.ts';
+import KapacitorAlert from '@/interfaces/Alert.ts';
 import DateUtil from '@/util/DateUtil.ts';
 import ColorUtil from '@/util/ColorUtil.ts';
